% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet_period.R
\name{dqp_sub_noreset}
\alias{dqp_sub_noreset}
\title{dqp_sub_noreset}
\usage{
dqp_sub_noreset(d, qp, included)
}
\arguments{
\item{d}{vector of \code{POSIX} event dates}

\item{qp}{quiet period}

\item{included}{indexes of elements of \code{d} already identified for
inclusion}
}
\value{
indices of elements in \code{d} that should be included based on
quiet period \code{qp}
}
\description{
recursive function for applying quiet period
}
\keyword{internal}
