% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet_period.R
\name{dqp_sub_reset}
\alias{dqp_sub_reset}
\title{dqp_sub_reset}
\usage{
dqp_sub_reset(d, qp)
}
\arguments{
\item{d}{vector of \code{POSIX} event dates}

\item{qp}{quiet period}
}
\value{
indices of elements in \code{d} that should be included based on
quiet period \code{qp}
}
\description{
function for applying quiet period that resets on subsequent events
}
\keyword{internal}
